# Speed up roc_auc_scorer by 25%