*Issue #, if available:*
#2318

*Description of changes:*
Follow-up to #2318, speed up by an additional 25% by only performing the index optimization logic if the number of rows is >100K.

Benchmark:

```python3
import numpy as np
import sklearn
from autogluon.core.metrics import roc_auc, make_scorer
import time

roc_auc_sklearn = make_scorer('roc_auc',
                      sklearn.metrics.roc_auc_score,
                      greater_is_better=True,
                      needs_threshold=True)

for sample_size in [
    10,
    100,
    1000,
    10000,
    100000,
]:
    y_true = np.random.randint(low=0, high=2, size=sample_size)
    y_score = np.random.random_sample(sample_size)

    print(f'### sample_size={sample_size} ###')
    ts = time.time()
    for i in range(1000):
        expected_score = roc_auc_sklearn(y_true, y_score)
    te = time.time()
    print(f'Sklearn: {round(te-ts, 2)}ms')

    ts = time.time()
    for i in range(1000):
        actual_score = roc_auc(y_true, y_score)
    te = time.time()
    print(f'Fast: {round(te-ts, 2)}ms')
    print(f'')
    assert np.isclose(expected_score, actual_score)

```

Out: (ML = mainline, PR = this PR version)

```
### sample_size=10 ###
Sklearn: 0.8ms
Fast (ML): 0.14ms
Fast (PR): 0.12ms
### sample_size=100 ###
Sklearn: 0.73ms
Fast (ML): 0.15ms
Fast (PR): 0.12ms
### sample_size=1000 ###
Sklearn: 0.91ms
Fast (ML): 0.26ms
Fast (PR): 0.21ms
### sample_size=10000 ###
Sklearn: 2.93ms
Fast (ML): 1.36ms
Fast (PR): 1.16ms
### sample_size=100000 ###
Sklearn: 28.33ms
Fast (ML): 15.74ms
Fast (PR): 14.36ms
```

By submitting this pull request, I confirm that you can use, modify, copy, and redistribute this contribution, under the terms of your choice.